Some refactoring

+++ lib/routes/upload_download.rb
@@ -1,23 +1,10 @@
 module WebSync
   module Routes
     class UploadDownload < Base
-      get '/upload' do
-        login_required
-        cache do
-          erb :upload
-        end
-      end
-      post '/upload' do
-        login_required
-        if params[:file]==nil
-          redirect "/upload"
-        end
-        tempfile = params[:file][:tempfile]
-        filename = params[:file][:filename]
-        filetype = params[:file][:type]
-        content = nil
-        # TODO: Split upload/download into its own external server. Right now Unoconv is blocking. Also issues may arise if multiple copies of LibreOffice are running on the same server. Should probably use a single server instance of LibreOffice
-        if params["convert"]
+
+      helpers do
+        def convert_file tempfile
+          filetype = MIME::Types.type_for(tempfile.path).first.content_type
           if filetype=="application/pdf"
             content = PDFToHTMLR::PdfFilePath.new(tempfile.path).convert.force_encoding("UTF-8")
           elsif filetype=='text/html'
@@ -32,6 +19,49 @@ module WebSync
               logger.info "Unoconv failed and Unrecognized filetype: #{params[:file][:type]}"
             end
           end
+          content
+        end
+        def sanitize_upload dom
+          # Basic security check
+          dom.css("script").remove();
+        end
+        def conversion_fail
+          flash[:danger] = "'#{h params[:file][:filename]}' failed to be converted."
+          redirect "/"
+        end
+        def upload_no_convert
+          file = params['file']
+          type = file[:type]
+          # Fingerprint file for mime-type if we aren't provided with it.
+          if type == 'application/octet-stream'
+            type = MIME::Types.type_for(file[:tempfile].path).first.content_type
+          end
+          blob = WSFile.create(name: file[:filename], content_type: type, edit_time: DateTime.now, create_time: DateTime.now)
+          blob.data = file[:tempfile].read
+          perm = Permission.create(user: current_user, file: blob, level: "owner")
+          flash[:success] = "'#{h file[:filename]}' was successfully uploaded."
+          redirect "/"
+        end
+      end
+
+      get '/upload' do
+        login_required
+        cache do
+          erb :upload
+        end
+      end
+
+      post '/upload' do
+        login_required
+        if params[:file]==nil
+          redirect "/upload"
+        end
+        tempfile = params[:file][:tempfile]
+        filename = params[:file][:filename]
+        content = nil
+        # TODO: Split upload/download into its own external server. Right now Unoconv is blocking. Also issues may arise if multiple copies of LibreOffice are running on the same server. Should probably use a single server instance of LibreOffice
+        if params["convert"]
+          content = convert_file tempfile
           File.delete tempfile.path
           if content!=nil
             dom = Nokogiri::HTML(content)
@@ -44,11 +74,12 @@ module WebSync
                 img["src"] = "assets/#{path}"
               end
             end
-            # Basic security check
-            dom.css("script").remove();
+            sanitize_upload dom
             doc = WSFile.create(
               name: filename,
-              body: {html: dom.to_html},
+              body: {
+                html: dom.to_html
+              },
               create_time: Time.now,
               edit_time: Time.now,
               content_type: 'text/websync'
@@ -56,6 +87,7 @@ module WebSync
             doc.assets = AssetGroup.get(1).assets
             doc.save
             perm = Permission.create(user: current_user, file: doc, level: "owner")
+
             # Upload images
             upload_list.each do |file|
               path = "/tmp/#{file}"
@@ -69,27 +101,16 @@ module WebSync
               flash[:success] = "'#{h params[:file][:filename]}' was successfully converted."
               redirect "/#{doc.id.encode62}/edit"
             else
-              flash[:danger] = "'#{h params[:file][:filename]}' failed to be converted."
-              redirect "/"
+              conversion_fail
             end
           else
-            flash[:danger] = "'#{h params[:file][:filename]}' failed to be converted."
-            redirect "/"
+            conversion_fail
           end
         else
-          file = params["file"]
-          type = file[:type]
-          # Fingerprint file for mime-type if we aren't provided with it.
-          if type=="application/octet-stream"
-            type = MIME::Types.type_for(file[:tempfile].path).first.content_type
-          end
-          blob = WSFile.create(name: file[:filename], content_type: type, edit_time: DateTime.now, create_time: DateTime.now)
-          blob.data = file[:tempfile].read
-          perm = Permission.create(user: current_user, file: blob, level: "owner")
-          flash[:success] = "'#{h file[:filename]}' was successfully uploaded."
-          redirect "/"
+          upload_no_convert
         end
       end
+
       # This doesn't need to verify authentication because the token is a 16 byte string.
       get '/:doc/download/:id' do
         doc = document_auth
